# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TestOps Companion API
  version: 2.9.0
  description: API specification for TestOps Companion — source of truth for frontend type generation.
servers:
  - url: /api/v1
    description: V1 API
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PipelineType:
      type: string
      enum: [jenkins, github-actions, custom]
    PipelineStatus:
      type: string
      enum: [success, failed, running, pending]
    TestRunStatus:
      type: string
      enum: [success, failed, running, pending, skipped, flaky]
...
    Pipeline:
      type: object
      required: [id, name, type, status, lastRun, successRate, config]
      properties:
        id: { type: string }
        name: { type: string }
        type: { $ref: '#/components/schemas/PipelineType' }
        status: { $ref: '#/components/schemas/PipelineStatus' }
        lastRun: { type: string }
        successRate: { type: number }
        config: { type: object, additionalProperties: true }
    CreatePipelineRequest:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { $ref: '#/components/schemas/PipelineType' }
        config: { type: object, additionalProperties: true }
    UpdatePipelineRequest:
      type: object
      properties:
        name: { type: string }
        type: { $ref: '#/components/schemas/PipelineType' }
        config: { type: object, additionalProperties: true }
        status: { type: string }

    # ── Test Run ──
    TestRun:
      type: object
      required: [id, pipelineId, pipelineName, status, startTime, endTime, duration, errorCount]
      properties:
        id: { type: string }
        pipelineId: { type: string }
        pipelineName: { type: string }
        status: { $ref: '#/components/schemas/TestRunStatus' }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        duration: { type: number }
        errorCount: { type: integer }
        screenshots: { type: array, items: { type: string } }
...
    FlakyTest:
      type: object
      properties:
        testName: { type: string }
        totalRuns: { type: integer }
        failureCount: { type: integer }
        flipFlopCount: { type: integer }
        flakinessScore: { type: number }
        severity: { type: string }
        lastFlakedAt: { type: string, format: date-time }

    # ── AI ──
...
paths:
...
  /pipelines:
    get:
      tags: [Pipelines]
      summary: List pipelines
      responses:
        '200':
          description: Pipeline list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Pipeline' }
    post:
      tags: [Pipelines]
      summary: Create pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePipelineRequest' }
      responses:
        '201':
          description: Pipeline created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pipeline' }
  /pipelines/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Pipelines]
      summary: Get pipeline by ID
      responses:
        '200':
          description: Pipeline details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pipeline' }
        '404':
          description: Pipeline not found
    put:
      tags: [Pipelines]
      summary: Update pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePipelineRequest' }
      responses:
        '200':
          description: Pipeline updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pipeline' }
    delete:
      tags: [Pipelines]
      summary: Delete pipeline
      responses:
        '204':
          description: Pipeline deleted
  /pipelines/{id}/start:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Pipelines]
      summary: Start pipeline
      responses:
        '200':
          description: Test run started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestRun' }
  /pipelines/{id}/test-runs:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Pipelines]
      summary: Get pipeline test runs
      responses:
        '200':
          description: Test runs for pipeline
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TestRun' }
  /pipelines/{id}/schedule:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Pipelines]
      summary: Schedule pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schedule]
              properties:
                schedule: { type: string }
      responses:
        '200':
          description: Pipeline scheduled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /pipelines/{id}/failed-tests:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Pipelines]
      summary: Get failed tests for pipeline
      responses:
        '200':
          description: Failed tests
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    testName: { type: string }
                    error: { type: string }
                    status: { type: string }
  /pipelines/{id}/flakey-tests:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Pipelines]
      summary: Get flaky tests for pipeline
      responses:
        '200':
          description: Flaky tests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FlakyTest' }
  /pipelines/validate-config:
    post:
      tags: [Pipelines]
      summary: Validate pipeline configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  errors:
                    type: array
                    items: { type: string }

  # ════════════════════════════════════════
  # TEST RUNS
  # ════════════════════════════════════════
...